requirements.txt
flask==2.3.3
supabase==1.0.4
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
